services/auth/src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
